introduce an id return value from the SQLiteDatabaseService when a climb is added and associated test

+++ BMG/Services/DatabaseService.swift
@@ -10,8 +10,8 @@ import Foundation
 
 protocol DatabaseService {
         
-    func addRockClimbToDb (routeName: String)  -> Int
+    func addRockClimbToDb (routeName: String)  -> Int64
     
-    func getRockClimbDataFromDb(idOfRockClimb: Int)
+    func getRockClimbDataFromDb(idOfRockClimb: Int64)
     
 }
